requirements.txt

flask==3.0.0
flask-login==0.6.3
flask-cors==4.0.0
werkzeug==3.0.1
pyodbc==5.0.1
pandas==2.1.4
openpyxl==3.1.2
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
pillow==10.1.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.2.2
ollama==0.1.6
python-dotenv==1.0.0
